simulator/requirements.txt
boto3==1.34.44
botocore==1.34.44
jmespath==1.0.1
numpy @ file:///private/tmp/numpy-20240108-5731-uciys2/numpy-1.26.3
packaging @ file:///private/tmp/python-packaging-20231211-10905-adsfgp/packaging-23.2
paho-mqtt==1.5.1
ply @ file:///private/tmp/python-ply-20231123-5320-9ahdwf/ply-3.11
protobuf==4.25.2
PyQt3D==5.15.6
PyQt5==5.15.10
PyQt5-sip==12.13.0
PyQtChart==5.15.6
PyQtDataVisualization==5.15.5
PyQtNetworkAuth==5.15.5
PyQtPurchasing==5.15.5
PyQtWebEngine==5.15.6
python-dateutil==2.8.2
s3transfer==0.10.0
sip @ file:///private/tmp/sip-20240124-4948-ufa6r6/sip-6.8.2
six==1.16.0
urllib3==2.0.7
py-dotenv==0.19.1
```
